Try to write my own minimal autopairs plugin
I don't want nothing fancy. I just never want a match to be inserted
when I don't want to. Try to be more restrictive than any other autpairs
plugin I have tried. Goal is to _never_ insert a matching tag if I don't
want it. Prefer to _not_ insert a match when I want to if that might
insert a match in a different scenario when I don't want